+++ eras/shelley/impl/CHANGELOG.md
@@ -14,6 +14,11 @@
 
 ### `testlib`
 
+* Add `resetAddrTxWits`, which re-derives key witnesses for a modified transaction body
+* Rename `freshBootstapAddress` to `freshBootstrapAddress` and express it in terms of the new
+  `freshBootstrapAddressWithPayloadSize`, which takes the HD payload size, or `Nothing` for a
+  plain Byron address
+* Add `freshBootstrapAddressOversizedPayload` and `largestBootstrapAddressAttrsSize`
 * Add `era` parameter to `PoolCert`s and `StakePoolParams`
 * Add `submitFailingSubsetTx{,M}`
 * Make `fixupTxOuts` parametric on level